Add option for download prefix in hosted dependencies
:Release Notes:
Implemented a new flag `-d, --download-prefix` to customize download
paths

:Detailed Notes:
A new command-line option `-d, --download-prefix` was added to allow
users to specify a custom prefix for the download directory of hosted
dependencies.
The README.md file was also updated to include usage information
about the new flag and an example showing how to use the flag for
specifying download directory prefixes.