Enhance handling and writing of git and hosted entries
:Release Notes:
Improve the processing of Git and hosted entries by adding checks for
duplicate URLs in Git entries and enhancing remote resolution.

:Detailed Notes:
For Git entries case, only the source bare repository is populated in
`git/cache`. This is automatically regenerated to the version entry
when the `pub get` command is executed.

Change-Id: I1dd3584f09b99001fe9e0e589db265a929105ab2